✨ Add isCommitted state
